@@ -5,12 +5,11 @@ namespace App\Http\Controllers\Api;
 
 
 use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-use Illuminate\Http\Response;
+use Illuminate\Http\JsonResponse;
 
 class ApiController extends Controller
 {
-    protected $perPage = 10;
+    protected int $perPageInit = 10;
 
     /**
      * @OA\Info(title="USEDESCK_TEST Api", version="1")
@@ -25,34 +24,57 @@ class ApiController extends Controller
      */
 
     /**
-     * @param Request $request
-     * @param $code
-     * @return \Illuminate\Http\JsonResponse
+     * @return JsonResponse
      */
-    public function sendError(Request $request, $code)
+    public function sendUnauthorized()
     {
-        return $this->sendResponse(false, $request->get('message', 'Error'), $code);
+        return $this->sendResponse('Unauthorized', JsonResponse::HTTP_UNAUTHORIZED);
     }
 
+    /**
+     * @param $model
+     * @return JsonResponse
+     */
     public function sendCreated($model)
     {
-        return $this->sendResponse(true, 'Success create', 201, $model);
+        return $this->sendResponse('Success create', JsonResponse::HTTP_CREATED, $model);
     }
 
-    public function sendSuccess($data)
+    /**
+     * @param $message
+     * @param array $data
+     * @return JsonResponse
+     */
+    public function sendSuccess($message, $data = [])
     {
-        return $this->sendResponse(true, $data, 200);
+        return $this->sendResponse($message, JsonResponse::HTTP_OK, $data);
     }
 
-    public function sendResponse($success, $message, $code, $data = [])
+    /**
+     * @return JsonResponse
+     */
+    public function notFound() {
+        return $this->sendResponse('Not found', JsonResponse::HTTP_NOT_FOUND, []);
+    }
+
+    /**
+     * @param $message
+     * @param $code
+     * @param array $data
+     * @return JsonResponse
+     */
+    public function sendResponse($message, $code, $data = [])
     {
         return response()->json(
             [
-                'success' => $success,
                 'message' => $message,
                 'data' => $data,
             ],
             $code
         );
+    }
+
+    public function getPerPage() {
+        return self::perPage;
     }
 }
