@@ -4,34 +4,34 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Requests\ClientRequest;
 use App\Http\Resources\ClientResource;
-use App\Models\Client;
+use App\Repositories\ClientRepository;
+use Exception;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
-use Illuminate\Http\Response;
 
 class ClientController extends ApiController
 {
-    private $model;
-    protected $perPage;
+    private ClientRepository $repository;
+    protected int $perPage;
 
-    public function __construct(Client $model)
+    public function __construct(ClientRepository $repository)
     {
-        $this->model = $model;
-        $this->perPage = request('per_page', null) ?: $this->perPage;
+        $this->repository = $repository;
+        $this->perPage = request('per_page', $this->perPageInit);
     }
 
     /**
-     * Список клиентов
+     * Clients list
      *
      * @OA\Get(
      *      path="/api/clients",
      *      operationId="getClientList",
      *      tags={"Clients"},
-     *      summary="Список клиентов",
+     *      summary="Clients list and search",
      *      security={ {"bearerAuth": {} } },
      *      @OA\Parameter(
      *          name="per_page",
-     *          description="Количество элементов на страницу",
+     *          description="Items per page",
      *          in="query",
      *          @OA\Schema(
      *              type="integer"
@@ -39,15 +39,31 @@ class ClientController extends ApiController
      *      ),
      *      @OA\Parameter(
      *          name="page",
-     *          description="Номер страницы",
+     *          description="Page number",
      *          in="query",
      *          @OA\Schema(
      *              type="integer"
      *          )
      *      ),
+     *     @OA\Parameter(
+     *          name="search_field",
+     *          description="Search field",
+     *          in="query",
+     *          @OA\Schema(
+     *              type="string"
+     *          )
+     *      ),
+     *     @OA\Parameter(
+     *          name="search",
+     *          description="Search value",
+     *          in="query",
+     *          @OA\Schema(
+     *              type="string"
+     *          )
+     *      ),
      *      @OA\Response(
      *         response=200,
-     *         description="successful operation",
+     *         description="Successful operation",
      *         @OA\JsonContent(
      *          @OA\Property(
      *              property="success",
@@ -123,11 +139,14 @@ class ClientController extends ApiController
      *     @OA\Response(response=400, description="Bad request"),
      * )
      *
-     * @return AnonymousResourceCollection|Response
+     * @param Request $request
+     * @return ClientResource
      */
-    public function index()
+    public function index(Request $request)
     {
-        return new ClientResource(Client::with(['emails', 'phones'])->paginate($this->perPage));
+        return new ClientResource($this->repository->all(
+            $this->perPage, $request->get('search_field'), $request->get('search')
+        ));
     }
 
     /**
@@ -135,13 +154,13 @@ class ClientController extends ApiController
      *
      * @OA\POST(
      *     path="/api/clients",
-     *     summary="Добавление клиента",
+     *     summary="Add client",
      *     tags={"Clients"},
      *     security={ {"bearerAuth": {} } },
-     *     @OA\Response(response="200", description="Успех"),
-     *     @OA\Response(response="400", description="Ошибка в запросе"),
-     *     @OA\Response(response="401", description="Ошибка авторизации"),
-     *     @OA\Response(response="422", description="Ошибка валидации"),
+     *     @OA\Response(response="200", description="Success"),
+     *     @OA\Response(response="400", description="Error in question"),
+     *     @OA\Response(response="401", description="Authorisation Error"),
+     *     @OA\Response(response="422", description="Validation error"),
      *     @OA\RequestBody(
      *        request="client",
      *        required=true,
@@ -154,54 +173,122 @@ class ClientController extends ApiController
      *)
      *
      * @param ClientRequest $request
-     * @return Response
+     * @return JsonResponse
      */
     public function store(ClientRequest $request)
     {
-        $client = Client::create($request->all());
-
-        if($phones = $request->get('phones', null)) {
-            $client->phones()->createMany($phones);
-        }
-
-        if($emails = $request->get('emails', null)) {
-            $client->emails()->createMany($emails);
-        }
-
-        return $this->sendCreated($client->with(['emails', 'phones']));
+        $client = $this->repository->create($request);
+        return $this->sendCreated($this->repository->find($client->id));
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
-     * @return Response
+     * @OA\GET(
+     *     path="/api/clients/{client_id}",
+     *     summary="Show client",
+     *     tags={"Clients"},
+     *     security={ {"bearerAuth": {} } },
+     *     @OA\Parameter(
+     *          name="client_id",
+     *          description="client id",
+     *          required=true,
+     *          in="path",
+     *          @OA\Schema(
+     *              type="integer"
+     *          )
+     *     ),
+     *     @OA\Response(response="200", description="Success"),
+     *     @OA\Response(response="400", description="Error in question"),
+     *   ),
+     *)
+     *
+     * @param int $id
+     * @return JsonResponse
      */
     public function show($id)
     {
-        //
+        if ($client = $this->repository->find($id)) {
+            return $this->sendSuccess('Client shown successfully', $client);
+        }
+        return $this->notFound();
     }
 
     /**
      * Update the specified resource in storage.
      *
+     * @OA\PUT(
+     *   path="/api/clients/{client_id}",
+     *   summary="Update client",
+     *   tags={"Clients"},
+     *   security={ {"bearerAuth": {} } },
+     *   @OA\Parameter(
+     *         name="client_id",
+     *         in="path",
+     *         required=true,
+     *         @OA\Schema(
+     *             type="integer",
+     *             format="int64",
+     *             minimum=1
+     *         )
+     *   ),
+     *   @OA\RequestBody(
+     *        request="Client",
+     *        required=true,
+     *        @OA\MediaType(
+     *            mediaType="application/json",
+     *            @OA\Schema(ref="#/components/schemas/Client")
+     *       )
+     *    ),
+     * @OA\Response(response="200", description="Client updated",),
+     * @OA\Response(response="400", description="Error in question"),
+     * @OA\Response(response="401", description="Authorisation Error"),
+     * )
+     *
      * @param Request $request
-     * @param  int  $id
-     * @return Response
+     * @param int $id
+     * @return JsonResponse
      */
-    public function update(ClientRequest $request, $id)
+    public function update(Request $request, $id)
     {
-        //
+        if ($this->repository->update($request, $id)) {
+            return $this->sendSuccess('Успешно обновлен', $this->repository->find($id));
+        }
+        return $this->notFound();
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
-     * @return Response
+     * @OA\DELETE(
+     *     path="/api/clients/{client_id}",
+     *     summary="Delete client",
+     *     tags={"Clients"},
+     *     security={ {"bearerAuth": {} } },
+     *     @OA\Parameter(
+     *          name="client_id",
+     *          description="client id",
+     *          required=true,
+     *          in="path",
+     *          @OA\Schema(
+     *              type="integer"
+     *          )
+     *     ),
+     *     @OA\Response(response="200", description="Success"),
+     *     @OA\Response(response="401", description="Authorisation Error"),
+     *   ),
+     *)
+     * @param int $id
+     * @return JsonResponse
+     * @throws Exception
      */
     public function destroy($id)
     {
-        //
+        $client = $this->repository->find($id);
+        if ($client) {
+            $client->delete();
+            return $this->sendSuccess('Client successfully deleted');
+        }
+        return $this->notFound();
     }
 }
